Disable data persistency
Currently the Redis default setting is used, which is that a periodic
snapshot writing to disk is enabled.

Its implementation is quite optimal, but as backup snapshots are not
needed for anything in current configuration, those are disabled in
this commit.

Redis server instance is defined to be an init process in DBaaS service
pod/container. In case of service restart the container will restart
also. There is no external storage for `/var/lib/redis/dump.rdb`
file and thus its content will be lost in service restart.

Redis data persistency over different failure scenarios is a future
feature candidate of RIC.

Change-Id: Ic9216bcb5067ed59c846e2b7805a78ee675bbc35
Signed-off-by: Rolf Badorek <rolf.badorek@nokia.com>