Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why is there neither a CMD nor ENTRYPOINT in the new Dockerfile?

Left them out as execution will happen from helm charts. Also the configuration generation was left out as that will also happen from helm charts. So the questions is, should I be able to use RIC dockerfiles as "standalone" or do we expect that certain kind of helm charts allways exists? This is a bit unclear for me.

Patch-set: 1
Label: Code-Review=0