Free resources when 'nget.noatomic' thread terminates
Call 'pthread_detach' from newly created thread.

Update unit tests to check that 'pthread_detach'
is called.

Previously the redismodule command 'nget.noatomic'
started a thread but the started thread was never
cleaned up because by default pthread cleans up
threads only when 'pthread_join' is called.
In case of 'nget.noatomic' the thread from which
'pthread_create' is called terminates immediately
and 'pthread_join' is never called resulting in
a resource leak.
We expect the created thread to be joined.
Therefore we need to call 'pthread_detach' from the
newly created thread which signals to pthread that
the resources allocated by the newly created thread
can be deallocated when the thread terminates.

Signed-off-by: Heinonen Arvo <arvo.heinonen@nokia.com>
Change-Id: If4d076fc4d58f65896638e99948c2d91a0000a9b